server/requirements.txt
flask
flask-sqlalchemy
flask-socketio
flask-session
flask-cors
sqlalchemy
sqlalchemy-utils
eventlet
gunicorn
pathlib
mysql-connector
python-dotenv
redis
bcrypt
pycryptodome